Class balanced sampler correctly calculates per hit class stats. train.py, using new sampler.
