+++ data/sampler.py
@@ -294,139 +294,3 @@ class BalancedGridGeoSampler(GeoSampler):
             int: Number of patches that will be sampled.
         """
         return self.length
-
-
-class ClassBalancedRandomBatchGeoSampler(BatchGeoSampler):
-    """Samples batches of elements from a region of interest randomly.
-
-    This is modified from RandomBatchGeoSampler so it can process bounding boxes
-    that are smaller than the given size and thus provide data balanced between
-    background and features. Note that randomly sampled chips may overlap.
-    """
-
-    def __init__(self, config) -> None:
-        """Initialize a new Sampler instance.
-
-        Args:
-            config: A dictionary containing the following keys:
-                - dataset: dataset to index from
-                - size: dimensions of each patch
-                - batch_size: number of samples per batch
-                - length: number of samples per epoch
-                - roi: region of interest to sample from
-                - units: defines if size is in pixel or CRS units
-        """
-        super().__init__(config["dataset"], config.get("roi"))
-        self.size = _to_tuple(config["size"])
-
-        if config.get("units") is None or config.get("units") == Units.PIXELS:
-            self.size = (self.size[0] * self.res, self.size[1] * self.res)
-
-        self.batch_size = config["batch_size"]
-        self.length = 0
-        self.hits, self.areas, self.class_areas = (
-            self.calculate_hits_and_areas()
-        )  # Added self.class_areas to store the per‐hit class area distributions
-
-        if config.get("length") is not None:
-            self.length = config["length"]
-
-    def calculate_hits_and_areas(
-        self,
-    ):  # Balancing (between background and features, i.e any class) happens here! We are creating different sized patches?!
-        """Calculate hits and areas for the dataset.
-
-        Returns:
-            hits: List of hits within the region of interest.
-            areas: Tensor of areas corresponding to each hit.
-        """
-        hits = []
-        areas = []
-        class_areas = []  # Adding a new list to keep track of class areas
-        context_x = self.size[1] / 2
-        context_y = self.size[0] / 2
-
-        for hit in self.index.intersection(tuple(self.roi), objects=True):
-            bounds = BoundingBox(*hit.bounds)
-            if (
-                bounds.maxx - bounds.minx >= self.size[1]
-                and bounds.maxy - bounds.miny >= self.size[0]
-            ):
-                shape_bounds = self.get_shape_bounds(bounds, context_x, context_y)
-                if (
-                    shape_bounds.maxx - shape_bounds.minx >= self.size[1]
-                    or shape_bounds.maxy - shape_bounds.miny >= self.size[0]
-                ):
-                    rows, cols = tile_to_chips(shape_bounds, self.size)
-                    self.length += rows * cols
-                else:
-                    self.length += 1
-                hits.append(hit)
-                areas.append(shape_bounds.area)
-
-                # Calling a new helper function compute_class_area()
-                per_class = {}  # dict: {class_id: area}
-                for class_id in config.KC_LABELS:  # KC_LABELS is defined in the config.py file, still need to find a way to code this to properly reference that dictionary
-                    # You may need to implement compute_class_area() using shapely or similar
-                    per_class[class_id] = compute_class_area(
-                        hit, class_id, shape_bounds
-                    )  # new helper function compute_class_area()
-                class_areas.append(per_class)
-
-        areas_tensor = torch.tensor(areas, dtype=torch.float)
-        if torch.sum(areas_tensor) == 0:
-            areas_tensor += 1
-
-        return hits, areas_tensor, class_areas
-
-    def get_shape_bounds(self, bounds, context_x, context_y):
-        """Get adjusted shape bounds considering context.
-
-        Args:
-            bounds: Original bounding box bounds.
-            context_x: Context size in x direction.
-            context_y: Context size in y direction.
-
-        Returns:
-            BoundingBox: Adjusted bounding box.
-        """
-        return BoundingBox(
-            bounds.minx + context_x,
-            bounds.maxx - context_x,
-            bounds.miny + context_y,
-            bounds.maxy - context_y,
-            bounds.mint,
-            bounds.maxt,
-        )
-
-    def __iter__(self):
-        """Return a batch of indices of a dataset.
-
-        Yields:
-            batch of (minx, maxx, miny, maxy, mint, maxt) coordinates to index a dataset
-        """
-        for _ in range(len(self)):
-            batch = []
-            batch_class_info = []  # list to keep track of the per batch class information
-            for _ in range(self.batch_size):
-                idx = torch.multinomial(self.areas, 1)
-                hit = self.hits[idx]
-                class_info = self.class_areas[
-                    idx
-                ]  # References the list crated in line 338. Might be better to output a dictionary with {class:area}
-                bounds = BoundingBox(*hit.bounds)
-                bounding_box = get_random_bounding_box(bounds, self.size, self.res)
-                batch.append(bounding_box)
-                batch_class_info.append(class_info)
-            yield {
-                "batch": batch,
-                "class_info": batch_class_info,
-            }  # Yielding the bach and the class info (dont know if this will break the dataloader's collate_fn when it calls the sampler)
-
-    def __len__(self):
-        """Returns the number of samples that this sampler will draw.
-
-        Returns:
-            int: The length of the sampler.
-        """
-        return self.length // self.batch_size
